chore: split benchmark
