+++ benchmark/fixture/strip_html.html
